Get rid ingest validation transformation and move it to clickhouse package
